Expected distributions of distances

These functions compute the expected distributions of distances between pairs of cases given a case reporting probability 'pi'. Analytical results are used for some special cases, including:

dpaircase(pi, alpha = 0.001)

dtemporal(pi, alpha = 0.001)

dspatial(pi, alpha = 0.001)

dgenetic(pi, alpha = 0.001)

dempiric(x, pi, alpha = 0.001)

Arguments

pi

The reporting probability, i.e. the proportion of cases of the outbreak that have been reported.

alpha

The probability threshold to be used to determine the maximum value of kappa; this value ('max_kappa') will be the smallest kappa so that p(kappa > max_kappa) < alpha. Defaults to 0.001.

x

A numeric vector providing the probability mass function, or empirical frequencies, of pairwise distances.

Details

temporal distances: the serial interval is assumed to be gamma-distributed; the procedure returns a discretized, weighted convolution of gamma distributions.

genetic distances: assumed ... TBC